feat(mapManager): mapManager module with assumption
assume 3 params in order; esri Map object, config settings, geoApi references
need to use the proper initialized geoApi to call the basemap.makeBasemaps(...)
todo: check to see if there is a way to call super or parent to access the main module without
the need to pass it.
tried require('./basemap.js'), the makeBasemaps() not working with require, not sure if some other
stuff is missing, or it just does not work due to the esriBundle not passed.