feat(ui): escape key closes panels in their opening order
rv-shell directive listens on the window scope instead of $rootScope since default is for focus to return to the HTML <body> tag when a focused element is destroyed. This caused issues with sidePanel that would not allow it to be closed on escape key. This is only a concern with multiple maps on the same page and the user is using keyboard navigation. The point information "mainDetails" panel cannot be closed without also closing the side panel due to how child panels are processed in stateManager (currently a childs parent is closed by default).

Closes #622