fix(ui): layer group visibility updates on child layer visibility change
Previously the visibility of a layer group did not correctly update when the visibility of a child layer changed. If all child layers for example were hidden, the group layer would remain visible. Now changes to child layer visibility correctly update the group layer. A group is considered visible if all children in the group are visible, otherwise it is hidden.

Closes #663