feat(ui): keyboard support and usability improvements for layer list panel
Previously the layer list only expanded on mouseover. Now the layer list expands when tabbing through items. Focus is also placed on the panels close button when tabbing off the first or last item, or when an item is selected with the keyboard. When using a mouse, the layer list is expanded with a slight delay which is triggered when the mouse movement speed is slowed over an item. jquery-hoverIntent library implements this logic.

Closes #147