fix(ui): do not set focus until transition is complete
The focus was being set to the close button on the main panel at the beginning of the transition (not waiting for child promises to resolve). The initial panel's position (shifted upwards) put the button above the app container (even above the top border of the page) and by focusing on it, the page would scroll up.

Closes #815