fix(ui): point info from ogcwms display in main panel
before: when we click on a feature coming from an OgcWMS the info
returned by the service doesn't display in the main panel. The link
function of the directive has no data to display.
solution: move the toggleDisplayPanel method into the then of the
promise
after: when we click on a feature coming from an ogcwms the info
returned by the service is displayed in the main panel.

Closes #1236