fix(test): add rv-langs with French as default in index-fgp-fr.html
Before: default language to English in index-fgp-fr.html . When there is
no rv-langs in the html, the default language setting is
["en-CA","fr-CA"] where English is at first place.
Solution: add rv-langs parameter = ["fr-CA","en-CA"] where French is at
first place. It is important to put also "en-CA" so the menu can display
that option under language section in menu.
After: default language to French in index-fgp-fr.html

Closes #1539