refactor(ui): reduce validation delay on layer import
There is a race condition between user pressing Enter and delayed change
update happening - if the user presses Enter too fast, latest changes to
the url might not be captured.

For example, copy-pasting
http://section917.cloudapp.net/arcgis/rest/services/TestData/aafc_crop_spatial_density_wheat/MapServer/, quickly adding 0, and pressing Enter,
will only capture the service url without its index, because of the
debounce delay. Reducing the delay, makes this less likely to happen