fix(ui): rvInteractive button widths adjusted
It seems datatables does not adjust the column width of `rvInteractive` based on the child width as would be expected. It could be using the literal value `rvInteractive` as a basis for the width.

It also seems it does not properly use the column width as defined in
our columnDefs.

These issues made it necessary to resort to manual width adjustments

Closes #2262